+++ src/main/java/riskyken/armourersWorkshop/client/render/tileEntity/RenderBlockMannequin.java
@@ -31,8 +31,8 @@ import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.ResourceLocation;
 import riskyken.armourersWorkshop.api.common.skin.data.ISkinDye;
 import riskyken.armourersWorkshop.api.common.skin.data.ISkinPointer;
-import riskyken.armourersWorkshop.client.gui.GuiMannequin;
-import riskyken.armourersWorkshop.client.gui.GuiMannequinTabSkinHair;
+import riskyken.armourersWorkshop.client.gui.mannequin.GuiMannequin;
+import riskyken.armourersWorkshop.client.gui.mannequin.GuiMannequinTabSkinHair;
 import riskyken.armourersWorkshop.client.model.ModelHelper;
 import riskyken.armourersWorkshop.client.model.ModelMannequin;
 import riskyken.armourersWorkshop.client.render.EntityTextureInfo;
